+++ django_glue/static/django_glue/js/django_glue_response.js
@@ -1,0 +1,4 @@
+class GlueResponse {
+    constructor() {
+    }
+}